Extract scheduleList code into own file
